Move the auth_active() method to device_utils
