+++ python/web/src/pi_cmds.py
@@ -5,7 +5,6 @@ Module for methods controlling and getting information about the Pi's Linux syst
 import subprocess
 import logging
 from flask_babel import _
-from settings import AUTH_GROUP
 
 
 def running_env():
@@ -121,19 +120,3 @@ def introspect_file(file_path, re_term):
         if match(re_term, line):
             return True
     return False
-
-
-def auth_active():
-    """
-    Inspects if the group defined in AUTH_GROUP exists on the system.
-    If it exists, tell the webapp to enable authentication.
-    Returns a (dict) with (bool) status and (str) msg
-    """
-    from grp import getgrall
-    groups = [g.gr_name for g in getgrall()]
-    if AUTH_GROUP in groups:
-        return {
-                "status": True,
-                "msg": _("You must log in to use this function"),
-                }
-    return {"status": False, "msg": ""}
